Initial code
The aim of this research is to apply different algorithms like K Nearest Neighbors, Support Vector Machine, Logistic Regression, Artificial Neural Network and to find the best method to classify the data. The dataset is found to be imbalanced as the ratio of no to yes in the response variable, default payment was 77.88:22.12. Hence, the number non-defaulters were observed to be approximately 3.5 times more than the number of defaulters. Hence SMOTE is used to balance the dataset. Also, significant features are selected for the better performance the models.